01 RAG For Articles/pyproject.toml
[project]
name = "RAG-Articles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "databricks-sdk>=0.20.0",
    "datasets==3.6.0",
    "feedparser>=6.0.12",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyter-dash>=0.4.2",
    "matplotlib>=3.10.6",
    "nltk>=3.9.3",
    "numpy>=2.3.3",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "plotly>=6.3.0",
    "protobuf==3.20.2",
    "psutil>=7.1.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "ragas>=0.3.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "sentence-transformers>=5.1.1",
    "setuptools>=80.9.0",
    "speedtest-cli>=2.1.3",
    "streamlit>=1.53.1",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.1",
]
[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
